+++ exporter.py
@@ -6,7 +6,7 @@ from threading import Thread
 from hca_ingest.api.ingestapi import IngestApi
 from kombu import Connection
 
-from exporter.graph.graph_crawler import GraphCrawler
+from exporter.graph.crawler import GraphCrawler
 from exporter.ingest.service import IngestService
 from exporter.metadata.service import MetadataService
 from exporter.queue.config import AmqpConnConfig, QueueConfig
