+++ tests/exporter/graph/test_graph_crawler.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 from hca_ingest.api.ingestapi import IngestApi
 from mock import MagicMock
 
-from exporter.graph.graph_crawler import GraphCrawler
+from exporter.graph.crawler import GraphCrawler
 from exporter.metadata.resource import MetadataResource
 from exporter.metadata.service import MetadataService
 from tests.mocks.files import MockEntityFiles
